+++ src/main/java/ru/yandex/practicum/filmorate/storage/dal/mappers/MpaRowMapper.java
@@ -13,7 +13,7 @@ public class MpaRowMapper implements RowMapper<Mpa> {
     public Mpa mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         Mpa mpa = new Mpa();
 
-        mpa.setId(resultSet.getLong("MPA_ID"));
+        mpa.setId(resultSet.getInt("MPA_ID"));
         mpa.setName(resultSet.getString("NAME"));
 
         return mpa;
