+++ src/main/java/ru/yandex/practicum/filmorate/storage/dto/UserDto.java
@@ -10,10 +10,10 @@ import java.util.Set;
 @Data
 public class UserDto {
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
-    private Long id;
+    private Integer id;
     private String email;
     private String login;
     private String name;
     private LocalDate birthday;
-    private Set<Long> friendsId = new HashSet<>();
+    private Set<Integer> friendsId = new HashSet<>();
 }
